Simplify selfUpdate()
This simplifies the selfUpdate() function by utilizing the
https://api.github.com/repos/xonixx/makesure/releases/latest
API call. The main benefit is avoiding having to guess the latest
version, but not running downloaded binaries automatially is a nice
bonus for those of us who'd like to inspect the updated makesure before
running it.

The other simplification here is relying on wget or curl for hanldling
HTTP errors instead of parsing them manually from the response. This was
done by using the '--fail-with-body' curl option, and removing the if
clause for wget that was ignoring the response code 8.

One test case was fixed to reflect new (somewhat more verbose) error
output.